requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
yfinance==0.2.38
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
xgboost==2.0.3
hmmlearn==0.3.2
arch==6.3.0
prometheus-client==0.20.0
psutil==5.9.8
mlflow==2.12.1
joblib==1.3.2
pyyaml==6.0.1
scipy==1.13.0
statsmodels==0.14.2
matplotlib==3.8.4
seaborn==0.13.2
requests==2.31.0
pydantic==2.6.4
python-multipart==0.0.9